refactor(fronend): incremental adoption of RRv7 middleware
