+++ frontend/app/@types/react-router.ts
@@ -1,16 +1,17 @@
-import 'react-router';
+import type { RouterContext } from 'react-router';
 
 import type { Namespace } from 'i18next';
 
-import type { ApplicationContext } from '~/.server/application-context';
+import type { ApplicationContext } from '~/.server/express/handlers';
 
 declare module 'react-router' {
-  ///
-  /// TODO ::: GjB ::: overriding RouterContextProvider here facilitates an incremental adoption of RRv7 middleware
-  ///                  obviously, this should be removed once the full migration to RRv7 middleware is complete
-  ///                  see: https://reactrouter.com/how-to/middleware#migration-from-apploadcontext
-  ///
-  interface RouterContextProvider extends ApplicationContext {}
+  // TODO ::: GjB ::: overriding RouterContextProvider here facilitates an incremental adoption of RRv7 middleware
+  //
+  // This should be removed once the full migration to RRv7 middleware is complete.
+  // see ~/.server/express/handlers for an explanation for why this is needed
+  interface RouterContextProvider extends ApplicationContext {
+    applicationContext: RouterContext<ApplicationContext>;
+  }
 
   /**
    * Route handles should export an i18n namespace, if necessary.
